server/requirements.txt
Django==3.2
psycopg2==2.9.1
django-nested-admin==3.3.3